refactor: group Open Graph module by feature dirs
