+++ projects/ngx-meta/src/open-graph/index.ts
@@ -1,26 +1,3 @@
-// Main providers
-export * from './src/ngx-meta-open-graph.module'
-export * from './src/ngx-meta-open-graph-profile.module'
-export * from './src/provide-ngx-meta-open-graph'
-export * from './src/provide-ngx-meta-open-graph-profile'
-// Types
-export * from './src/open-graph'
-export * from './src/open-graph-image'
-export * from './src/open-graph-metadata'
-export * from './src/open-graph-profile'
-export * from './src/open-graph-profile-gender'
-export * from './src/open-graph-type'
-// Specific metadata managers
-// Basic
-export * from './src/open-graph-title-metadata-provider'
-export * from './src/open-graph-type-metadata-provider'
-export * from './src/open-graph-image-metadata-provider'
-export * from './src/open-graph-url-metadata-provider'
-export * from './src/open-graph-description-metadata-provider'
-export * from './src/open-graph-locale-metadata-provider'
-export * from './src/open-graph-site-name-metadata-provider'
-// Profile
-export * from './src/open-graph-profile-first-name-metadata-provider'
-export * from './src/open-graph-profile-last-name-metadata-provider'
-export * from './src/open-graph-profile-username-metadata-provider'
-export * from './src/open-graph-profile-gender-metadata-provider'
+export { OpenGraphMetadata } from './src/open-graph-metadata'
+export * from './src/basic-optional'
+export * from './src/profile'
