+++ projects/ngx-meta/src/open-graph/src/basic-optional/index.ts
@@ -1,0 +1,2 @@
+export * from './managers'
+export * from './providers'
