+++ projects/ngx-meta/src/open-graph/src/basic-optional/managers/index.ts
@@ -1,0 +1,10 @@
+export { OpenGraph } from './open-graph'
+export { OPEN_GRAPH_DESCRIPTION_METADATA_PROVIDER } from './open-graph-description-metadata-provider'
+export { OPEN_GRAPH_IMAGE_METADATA_PROVIDER } from './open-graph-image-metadata-provider'
+export { OpenGraphImage } from './open-graph-image'
+export { OPEN_GRAPH_LOCALE_METADATA_PROVIDER } from './open-graph-locale-metadata-provider'
+export { OPEN_GRAPH_SITE_NAME_METADATA_PROVIDER } from './open-graph-site-name-metadata-provider'
+export { OPEN_GRAPH_TITLE_METADATA_PROVIDER } from './open-graph-title-metadata-provider'
+export { OPEN_GRAPH_TYPE_METADATA_PROVIDER } from './open-graph-type-metadata-provider'
+export * from './open-graph-type'
+export { OPEN_GRAPH_URL_METADATA_PROVIDER } from './open-graph-url-metadata-provider'
