+++ projects/ngx-meta/src/open-graph/src/basic-optional/managers/open-graph-image.ts
