+++ projects/ngx-meta/src/open-graph/src/basic-optional/managers/open-graph-site-name-metadata-provider.ts
@@ -1,4 +1,4 @@
-import { makeOpenGraphMetadataProvider } from './make-open-graph-metadata-provider'
+import { makeOpenGraphMetadataProvider } from '../../utils/make-open-graph-metadata-provider'
 import { _GLOBAL_APPLICATION_NAME } from '@davidlj95/ngx-meta/core'
 
 /**
