+++ projects/ngx-meta/src/open-graph/src/basic-optional/managers/open-graph-type.ts
