+++ projects/ngx-meta/src/open-graph/src/basic-optional/providers/index.ts
@@ -1,0 +1,2 @@
+export { NgxMetaOpenGraphModule } from './ngx-meta-open-graph.module'
+export { provideNgxMetaOpenGraph } from './provide-ngx-meta-open-graph'
