+++ projects/ngx-meta/src/open-graph/src/profile/index.ts
@@ -1,0 +1,2 @@
+export * from './managers'
+export * from './providers'
