+++ projects/ngx-meta/src/open-graph/src/profile/managers/index.ts
@@ -1,0 +1,10 @@
+export { OpenGraphProfile } from './open-graph-profile'
+export { OPEN_GRAPH_PROFILE_FIRST_NAME_METADATA_PROVIDER } from './open-graph-profile-first-name-metadata-provider'
+export { OPEN_GRAPH_PROFILE_LAST_NAME_METADATA_PROVIDER } from './open-graph-profile-last-name-metadata-provider'
+export { OPEN_GRAPH_PROFILE_GENDER_METADATA_PROVIDER } from './open-graph-profile-gender-metadata-provider'
+export {
+  OPEN_GRAPH_PROFILE_GENDER_MALE,
+  OPEN_GRAPH_PROFILE_GENDER_FEMALE,
+  OpenGraphProfileGender,
+} from './open-graph-profile-gender'
+export { OPEN_GRAPH_PROFILE_USERNAME_METADATA_PROVIDER } from './open-graph-profile-username-metadata-provider'
