+++ projects/ngx-meta/src/open-graph/src/profile/managers/open-graph-profile-gender-metadata-provider.ts
@@ -1,4 +1,4 @@
-import { makeOpenGraphProfileMetadataProvider } from './make-open-graph-profile-metadata-provider'
+import { makeOpenGraphProfileMetadataProvider } from '../utils/make-open-graph-profile-metadata-provider'
 
 /**
  * Manages the {@link OpenGraphProfile.gender} metadata
