+++ projects/ngx-meta/src/open-graph/src/profile/providers/index.ts
@@ -1,0 +1,2 @@
+export { NgxMetaOpenGraphProfileModule } from './ngx-meta-open-graph-profile.module'
+export { provideNgxMetaOpenGraphProfile } from './provide-ngx-meta-open-graph-profile'
