+++ projects/ngx-meta/src/open-graph/src/profile/providers/provide-ngx-meta-open-graph-profile.ts
@@ -1,8 +1,8 @@
 import { Provider } from '@angular/core'
-import { OPEN_GRAPH_PROFILE_FIRST_NAME_METADATA_PROVIDER } from './open-graph-profile-first-name-metadata-provider'
-import { OPEN_GRAPH_PROFILE_LAST_NAME_METADATA_PROVIDER } from './open-graph-profile-last-name-metadata-provider'
-import { OPEN_GRAPH_PROFILE_USERNAME_METADATA_PROVIDER } from './open-graph-profile-username-metadata-provider'
-import { OPEN_GRAPH_PROFILE_GENDER_METADATA_PROVIDER } from './open-graph-profile-gender-metadata-provider'
+import { OPEN_GRAPH_PROFILE_FIRST_NAME_METADATA_PROVIDER } from '../managers/open-graph-profile-first-name-metadata-provider'
+import { OPEN_GRAPH_PROFILE_LAST_NAME_METADATA_PROVIDER } from '../managers/open-graph-profile-last-name-metadata-provider'
+import { OPEN_GRAPH_PROFILE_USERNAME_METADATA_PROVIDER } from '../managers/open-graph-profile-username-metadata-provider'
+import { OPEN_GRAPH_PROFILE_GENDER_METADATA_PROVIDER } from '../managers/open-graph-profile-gender-metadata-provider'
 
 /**
  * Provides {@link https://ngx-meta.dev/built-in-modules/open-graph/ | Open Graph module}
